requirements.txt
pandas==2.1.4
opencv-python==4.9.0.80
numpy==1.26.2
scikit-learn==1.3.2
torch==2.1.2
transformers==4.36.2
Pillow==10.1.0